Initial code
Adds a tool that patches a reference FW image with a new Rust DXE
Core.

This script is meant to be focused and fast for patching specific
reference firmware images with a new Rust DXE Core. It currently
can patch a reference firmware in about half a second with minimal
dependencies on other (bloated) tools.

It is not meant to be a general-purpose firmware patching tool.

The tool is currently written to solve two specific problems:

1. Patching QEMU FW quickly for fast dev and test cycles on QEMU.
2. Patching Intel firmware quickly for fast dev and test cycles,
   but also so those without access to Intel firmware source code
   can patch Intel ROMs and test their changes on physical Intel
   platforms.

This is a very initial rough draft that has a QEMU Q35 config.

Future changes will add support for specific Intel platforms and
SBSA.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>